+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Adds latest `sentry-sdk[flask]` as an install dependency, and update Sentry logic to be able to send environment, app version and profiling/performance info [#3086](https://github.com/opendatateam/udata/pull/3086)
+- Add report get endpoint [#3115](https://github.com/opendatateam/udata/pull/3115)
 
 ## 9.1.2 (2024-07-29)